src/requirements.txt
Adafruit-GPIO==1.0.3
Adafruit-PureIO==1.1.11
Adafruit-SSD1306==1.6.2
arrow==0.17.0
flask==3.1.0
folium==0.12.0
gpiozero==1.6.2
h3==3.7.7
jinja2==3.1.6
loguru==0.7.3
logzero==1.7.0
pytz==2025.1
redis==5.2.1
requests==2.25.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.0
simplejpeg==1.6.4
spidev==3.5
suntime==1.3.2
timezonefinder==6.5.3
uWSGI==2.0.28
wget==3.2
xmltodict==0.14.2
